feat: Add separate nodeInfoService for nova
