+++ client/src/features/visualizer-threejs/VisualizerInstance.tsx
@@ -15,7 +15,6 @@ import { VisualizerRouteProps } from "../../app/routes/VisualizerRouteProps";
 import { ServiceFactory } from "../../factories/serviceFactory";
 import { useNetworkConfig } from "../../helpers/hooks/useNetworkConfig";
 import { IFeedBlockData } from "../../models/api/stardust/feed/IFeedBlockData";
-import { StardustFeedClient } from "../../services/stardust/stardustFeedClient";
 import { NovaFeedClient } from "../../services/nova/novaFeedClient";
 import { Wrapper } from "./wrapper/Wrapper";
 import "./Visualizer.scss";
@@ -58,7 +57,7 @@ const VisualizerInstance: React.FC<RouteComponentProps<VisualizerRouteProps>> = 
     const indexToBlockId = useTangleStore(s => s.indexToBlockId);
 
     const emitterRef = useRef<THREE.Mesh>(null);
-    const feedServiceRef = useRef<StardustFeedClient | null>(null);
+    const feedServiceRef = useRef<NovaFeedClient | null>(null);
 
     /**
      * Pause on tab or window change
@@ -172,7 +171,7 @@ const VisualizerInstance: React.FC<RouteComponentProps<VisualizerRouteProps>> = 
         if (!runListeners) {
             return;
         }
-        feedServiceRef.current = ServiceFactory.get<StardustFeedClient>(
+        feedServiceRef.current = ServiceFactory.get<NovaFeedClient>(
             `feed-${network}`
         );
         setIsPlaying(true);
