+++ client/src/services/stardust/nodeInfoService.ts
@@ -1,11 +1,9 @@
 import { INodeInfoBaseToken, IRent } from "@iota/sdk-wasm/web";
-import { StardustApiClient } from "./stardust/stardustApiClient";
+import { StardustApiClient } from "../stardust/stardustApiClient";
 import { ServiceFactory } from "~factories/serviceFactory";
 import { INodeInfoResponse as IStardustInfoResponse } from "~models/api/stardust/INodeInfoResponse";
-import { INodeInfoResponse as INovaInfoResponse } from "~models/api/nova/INodeInfoResponse";
-import { NOVA, STARDUST } from "~models/config/protocolVersion";
+import { STARDUST } from "~models/config/protocolVersion";
 import { NetworkService } from "~services/networkService";
-import { NovaApiClient } from "./nova/novaApiClient";
 
 /**
  * The reduced stardust node info fields relevant for Explorer.
@@ -30,20 +28,20 @@ export interface IStardustNodeInfo {
 }
 
 /**
- * Service to handle base token info.
+ * Service to handle base token info on stardust.
  */
 export class NodeInfoService {
     /**
      * Cache of the base taken infos.
      */
-    private _cache: { [network: string]: IStardustNodeInfo | INovaInfoResponse } = {};
+    private _cache: { [network: string]: IStardustNodeInfo } = {};
 
     /**
      * Get the base token info by network.
      * @param network The name of the network.
      * @returns The base token info.
      */
-    public get(network: string): IStardustNodeInfo | INovaInfoResponse {
+    public get(network: string): IStardustNodeInfo {
         return this._cache[network];
     }
 
@@ -53,7 +51,6 @@ export class NodeInfoService {
     public async buildCache(): Promise<void> {
         const networksService = ServiceFactory.get<NetworkService>("network");
         const stardustNetworks = networksService.networks().filter(n => n.protocolVersion === STARDUST);
-        const novaNetworks = networksService.networks().filter(n => n.protocolVersion === NOVA);
 
         for (const networkDetails of stardustNetworks) {
             const apiClient = ServiceFactory.get<StardustApiClient>(`api-client-${STARDUST}`);
@@ -63,18 +60,6 @@ export class NodeInfoService {
 
             if (baseToken && protocolVersion && bech32Hrp && rentStructure) {
                 this._cache[network] = { baseToken, protocolVersion, bech32Hrp, rentStructure };
-            }
-        }
-
-        // TODO Split this into separate services
-        for (const networkDetails of novaNetworks) {
-            const apiClient = ServiceFactory.get<NovaApiClient>(`api-client-${NOVA}`);
-            const network = networkDetails.network;
-            const response: INovaInfoResponse = await apiClient.nodeInfo({ network });
-            const { baseToken, status, protocolParameters } = response;
-
-            if (baseToken && status && protocolParameters) {
-                this._cache[network] = response;
             }
         }
     }
